Update Orieentiring_GA.ipynb
Added a dictionary in the Individual class that save the index (number of steps) in which it reaches a Control Point. (these change will be added to the Individual.py file as well)
Changed algorithms to retrieve the path to a Control Point, the path from a Control Point to the end and the path between two Control Points.